Separated existing functions to files
